+++ src/components/Kpi/DrawKpis.js
