+++ src/components/PrefectureTable/index.js
@@ -1,3 +1,3 @@
-import { drawPrefectureTable } from "./prefectureTable";
+import { drawPrefectureTable } from "./PrefectureTable";
 
 export default drawPrefectureTable;
