+++ src/components/SpreadTrendChart/SpreadTrend.js
